[batch] extract service backend test fixtures to separate file
Next PR will split service backend tests for better legibility so fixtures need to be in one place.
